📝 Add README to the Project
Added a README to the project, outlining a task-based C++ learning
process, as well as a series of tasks applying C++ to computer vision
and image processing. These tasks are structured to progressively
advance from beginner to intermediate and then to advanced levels
,increasing in difficulty. They are also designed to reinforce
foundational concepts in image processing and computer vision.